backend/pyproject.toml
[project]
name = "bi-agent-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.24.0",
    "fastapi>=0.115.0",
    "keyring>=25.7.0",
    "langchain>=1.2.0",
    "langchain-azure-ai>=1.1.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.4",
    "openai>=2.13.0",
    "requests>=2.33.1",
    "duckdb>=1.4.3",
    "airbyte>=0.44.1",
    "langgraph>=1.1.4",
    "langchain-anthropic==1.4.0",
    "pydantic[email]>=2.12.5",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.28.0",
    "sympy>=1.14.0",
    "uvicorn>=0.34.0",
    "gunicorn>=22.0.0",
    "plotly>=6.6.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
